fixed support for prt-sim and moved utils to common. removed unused docs and fixed tests. #minor
