--- docs/source/dynamic-programming.md
@@ -1,11 +1,0 @@
-# Dynamic Programming
-
-Dynamic programming is a method of solving an MDP using the Bellman Equations. It requires the entire transition
-function to be known.
-
-:::{mermaid}
-graph TD
-    A[Start] --> B{Is it working?}
-    B -->|Yes| C[Great!]
-    B -->|No| D[Fix it!]
-:::
