+++ src/prt_rl/common/__init__.py
@@ -1,0 +1,1 @@
+"""Common functions that are not specific to a single RL algorithm"""
