+++ src/prt_rl/common/policy/keyboard.py
@@ -3,7 +3,7 @@ import threading
 import torch
 from typing import Dict
 from prt_rl.env.interface import EnvParams
-from prt_rl.utils.policy.policies import Policy
+from prt_rl.common.policy.policies import Policy
 
 class KeyboardPolicy(Policy):
     """
