+++ src/prt_rl/exact/qlearning.py
@@ -2,12 +2,12 @@ from tensordict.tensordict import TensorDict
 import torch
 from typing import Optional, List, Any
 from prt_rl.env.interface import EnvironmentInterface
-from prt_rl.utils.loggers import Logger
-from prt_rl.utils.trainers import TDTrainer
-from prt_rl.utils.decision_functions import DecisionFunction
-from prt_rl.utils.policy import QTablePolicy
-from prt_rl.utils.schedulers import ParameterScheduler
-from prt_rl.utils.metrics import MetricTracker
+from prt_rl.common.loggers import Logger
+from prt_rl.common.trainers import TDTrainer
+from prt_rl.common.decision_functions import DecisionFunction
+from prt_rl.common.policy import QTablePolicy
+from prt_rl.common.schedulers import ParameterScheduler
+from prt_rl.common.metrics import MetricTracker
 
 class QLearning(TDTrainer):
     r"""
