+++ src/prt_rl/ppo.py
@@ -9,12 +9,12 @@ import torch.nn.functional as F
 from tensordict.tensordict import TensorDict
 from typing import Optional, List, Any
 from prt_rl.env.interface import EnvParams, EnvironmentInterface
-from prt_rl.utils.policy import ActorCriticPolicy
-from prt_rl.utils.loggers import Logger
-from prt_rl.utils.schedulers import ParameterScheduler
-from prt_rl.utils.progress_bar import ProgressBar
-from prt_rl.utils.metrics import MetricTracker
-from prt_rl.utils.trainers import ActorCriticTrainer
+from prt_rl.common.policy import ActorCriticPolicy
+from prt_rl.common.loggers import Logger
+from prt_rl.common.schedulers import ParameterScheduler
+from prt_rl.common.progress_bar import ProgressBar
+from prt_rl.common.metrics import MetricTracker
+from prt_rl.common.trainers import ActorCriticTrainer
 
 class PPO(ActorCriticTrainer):
     """
