--- src/prt_rl/utils/__init__.py
@@ -1,1 +1,0 @@
-"""Utility functions that are not specific to a single RL algorithm"""
