+++ tests/common/policy/test_qtable_policy.py
@@ -1,7 +1,7 @@
 from tensordict.tensordict import TensorDict
 import torch
 from prt_rl.env.interface import EnvParams
-from prt_rl.utils.policy import QTablePolicy
+from prt_rl.common.policy import QTablePolicy
 
 def test_qtable_policy():
     # Create a fake environment that has 1 discrete action [0,...,3] and 1 discrete state with the same interval
